backend/flask_app/requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
bcrypt==4.0.1
requests==2.31.0
python-barcode==0.14.0
Pillow==10.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7